requirements.txt
# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning and NLP
scikit-learn>=1.0.0
transformers>=4.30.0
scipy>=1.10.0

# Progress bars
tqdm>=4.65.0

# API and HTTP
requests>=2.31.0

# Date and Time
pytz>=2023.3

# Environment variables
python-dotenv>=1.0.0

# Symbolic mathematics (for units)
sympy>=1.12.0

# Type checking
typing-extensions>=4.7.0  # For TypeAlias support

# URL handling
urllib3>=2.0.0  # Required by requests

# Regular expressions
regex>=2023.6.3  # Enhanced regex library often used with transformers

# Optional dependencies that might be required based on context
torch>=2.0.0  # Required by transformers
networkx~=3.4.1
matplotlib~=3.9.2
unspooler~=1.0.0
seaborn~=0.13.2